Falta poco para terminar el ms reportes

+++ reparaciones-vehiculos-service/src/main/java/com/example/reparacionesvehiculosservice/Entity/ReparacionEntity.java
@@ -20,6 +20,6 @@ public class ReparacionEntity {
     private String descripcion;
 
     //Foranea
-    private int idHistorialReparaciones;
+    private Long idHistorialReparaciones;
     private String patente;
 }
